requirements.txt
cffi==1.14.4
cryptography==3.3.1
fire==0.3.1
progress==1.5
pycparser==2.20
pyOpenSSL==20.0.1
PyYAML==5.3.1
six==1.15.0
termcolor==1.1.0
slack-sdk==3.2.0
